Continuation of fix.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@18566 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/event.cpp
@@ -43,7 +43,7 @@ void csKeyEventHelper::GetModifiers (const iEvent* event,
   memset (&modifiers, 0, sizeof (modifiers));
 
   const void* mod;
-  uint32 modSize;
+  size_t modSize;
   if (event->Retrieve ("keyModifiers", mod, modSize) != csEventErrNone) 
     return;
   memcpy (&modifiers, mod, MIN (sizeof (modifiers), modSize));
